refactor(https_client): share the exclusive-temp-file primitive
Extracts TempSpoolFactory's atomic O_EXCL/owner-only-permissions temp
file creation (symlink-attack resistant) into exclusiveTempFile.{hpp,cpp},
so the upcoming file-backed compressor doesn't reimplement the same
security-sensitive logic a second time. Also folds in a fallback to
TMPDIR/TEMP/TMP/"/tmp" when the caller's directory is empty, since nothing
today actually populates ModuleConfig::spoolDir.

No behavior change for TempSpoolFactory itself.
